requirements.txt
fastapi[all]
openai
python-dotenv
aiofiles